chore(search-filter): moved filter component files to their own folder
